+++ src/app/(population-trends)/page.tsx
@@ -1,4 +1,4 @@
-import PopulationTrendsPage from "@/pages/PopulationTrendsPage";
+import PopulationTrendsPage from "@/feature/populationTrends/pages/PopulationTrendsPage";
 
 export default function Page() {
   return <PopulationTrendsPage />;
