+++ src/feature/populationTrends/components/LoadingPlaceholder/index.tsx
