+++ src/feature/populationTrends/components/PopulationChart/index.tsx
