+++ src/shared/components/Icon/Icon.test.tsx
